:goal_net: Enhanced Error Handling System for Product Service (#85)
* wip wip wip wip wip wip wip wip wip wip wip wip wip wip wip wip wip wip

* wip

* wip

* wip

* wip

* basic error

* Error handling update and delete category

* :art: move capitalizing of category name

* small adjustment in category service

* product exception handling

* :fire: remove

* :fire: unused

---------

+++ product-service/src/main/java/solutions/skydev/pos/product_service/service/CategoryServiceImpl.java
@@ -4,6 +4,9 @@ import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import solutions.skydev.pos.common.error.domain.ConflictException;
+import solutions.skydev.pos.common.error.domain.ResourceNotFoundException;
+import solutions.skydev.pos.common.error.domain.ValidationException;
 import solutions.skydev.pos.product_service.model.entity.Category;
 import solutions.skydev.pos.product_service.model.entity.Product;
 import solutions.skydev.pos.product_service.repository.CategoryRepository;
@@ -26,6 +29,27 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public Category addCategory(Category category) {
+        if (category.getName() == null || category.getName().isEmpty()) {
+            throw new ValidationException("Category name cannot be empty");
+        }
+        
+        // Capitalize the first letter of the category name
+        String capitalizedCategoryName = category.getName().substring(0, 1).toUpperCase() + category.getName().substring(1);
+        category.setName(capitalizedCategoryName);
+
+        if (category.getName().length() > 50 || category.getName().length() < 3) {
+            throw new ValidationException("Category name must be between 3 and 50 characters long");
+        }
+
+        if (categoryRepository.existsByName((category.getName()))) {
+            throw new ConflictException("Category with name '" + category.getName() + "' already exists");
+        }
+
+        // Should not container numbers or special characters
+        if (!category.getName().matches("^[a-zA-Z ]+$")) {
+            throw new ValidationException("Category name can only contain letters and spaces");
+        }
+
         return categoryRepository.save(category);
     }
 
@@ -38,14 +62,45 @@ public class CategoryServiceImpl implements CategoryService {
     public Category updateCategory(Long id, Category updatedCategory) {
         Category existingCategory = categoryRepository.findById(id).orElse(null);
         if (existingCategory == null) {
-            // TODO: handle the case where the product doesn't exist.
-            return null;
+            throw new ResourceNotFoundException("Category with id " + id + " not found");
         }
+
+        if (updatedCategory.getName() == null || updatedCategory.getName().isEmpty()) {
+            throw new ValidationException("Category name cannot be empty");
+        }
+        
+        String capitalizedCategoryName = updatedCategory.getName().substring(0, 1).toUpperCase() + updatedCategory.getName().substring(1);
+        updatedCategory.setName(capitalizedCategoryName);
+
+
+        if (updatedCategory.getName().length() > 50 || updatedCategory.getName().length() < 3) {
+            throw new ValidationException("Category name must be between 3 and 50 characters long");
+        }
+
+        if (categoryRepository.existsByName(updatedCategory.getName()) && !updatedCategory.getName().equals(existingCategory.getName())) {
+            throw new ConflictException("Category with name '" + updatedCategory.getName() + "' already exists");
+        }
+
+        if (!updatedCategory.getName().matches("^[a-zA-Z ]+$")) {
+            throw new ValidationException("Category name can only contain letters and spaces");
+        }
+
         return categoryRepository.save(updatedCategory);
     }
 
     @Override
+    @Transactional
     public void deleteCategory(Long id) {
+        Category existingCategory = categoryRepository.findById(id).orElse(null);
+        if (existingCategory == null) {
+            throw new ResourceNotFoundException("Category with id " + id + " not found");
+        }
+
+        // Check if the category has products associated with it
+        if (!existingCategory.getProducts().isEmpty()) {
+            throw new ConflictException("Category with id " + id + " cannot be deleted because it has products associated with it");
+        }
+
         categoryRepository.deleteById(id);
     }
